Move `backward` to the new `backward` package. Puts helper functions in `_utils.py`
