+++ src/torchjd/backward/__init__.py
@@ -1,0 +1,2 @@
+from torchjd.backward.backward import backward
+from torchjd.backward.multi_task_backward import multi_task_backward
