+++ src/torchjd/backward/backward.py
@@ -2,9 +2,15 @@ from typing import Iterable, Sequence
 
 from torch import Tensor
 
-from ._transform import Diagonalize, EmptyTensorDict, Init, Jac, Store, make_aggregation
-from ._transform.strategy import UnifyingStrategy
-from .aggregation import Aggregator
+from torchjd._transform import Diagonalize, EmptyTensorDict, Init, Jac, Store, make_aggregation
+from torchjd._transform.strategy import UnifyingStrategy
+from torchjd.aggregation import Aggregator
+
+from ._utils import (
+    _as_tensor_list,
+    _check_optional_positive_chunk_size,
+    _check_retain_graph_compatible_with_chunk_size,
+)
 
 
 def backward(
@@ -72,7 +78,7 @@ def backward(
     diag = Diagonalize(tensors)
 
     # Transform that computes the required jacobians
-    jac = Jac(tensors, inputs, chunk_size=parallel_chunk_size, retain_graph=retain_graph)
+    jac = Jac(tensors, inputs, parallel_chunk_size, retain_graph)
 
     # Transform that defines the aggregation of the jacobians into gradients
     aggregation = make_aggregation(UnifyingStrategy(A, inputs))
@@ -83,32 +89,3 @@ def backward(
     backward_transform = store << aggregation << jac << diag << init
 
     backward_transform(EmptyTensorDict())
-
-
-def _check_optional_positive_chunk_size(parallel_chunk_size: int | None) -> None:
-    if not (parallel_chunk_size is None or parallel_chunk_size > 0):
-        raise ValueError(
-            "`parallel_chunk_size` should be `None` or greater than `0`. (got "
-            f"{parallel_chunk_size})"
-        )
-
-
-def _as_tensor_list(tensors: Sequence[Tensor] | Tensor) -> list[Tensor]:
-    if isinstance(tensors, Tensor):
-        output = [tensors]
-    else:
-        output = tensors
-    return output
-
-
-def _check_retain_graph_compatible_with_chunk_size(
-    tensors: list[Tensor],
-    retain_graph: bool,
-    parallel_chunk_size: int | None,
-) -> None:
-    tensors_numel = sum([tensor.numel() for tensor in tensors])
-    if parallel_chunk_size is not None and parallel_chunk_size < tensors_numel and not retain_graph:
-        raise ValueError(
-            "When using `retain_graph=False`, parameter `parallel_chunk_size` must be `None` or "
-            "large enough to compute all gradients in parallel."
-        )
